// src/types/hrmv2/hrmv2.enum.ts
export enum EmployeeStatus {
  Working = 1,
  Pausing = 2,
  Quit = 3,
  MaternityLeave = 4,
}

export enum Sex {
  Male = 1,
  Female = 2,
}

export enum UserType {
  Internship = 0,
  Collaborators = 1,
  Staff = 2,
  ProbationaryStaff = 3,
  Vendor = 5,
}

export enum JobPosition {
  IT = 'IT',
  HR = 'HR',
}